requirements.txt
aniso8601==9.0.1
autopep8==1.5.7
bcrypt==3.2.0
blinker==1.4
bson==0.5.10
cffi==1.14.6
click==8.0.1
colorama==0.4.6
contourpy==1.0.6
cryptography==35.0.0
cycler==0.11.0
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==0.15.1
fonttools==4.38.0
greenlet==1.1.2
gunicorn==20.0.4
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.4.4
MarkupSafe==2.0.1
matplotlib==3.5.3
mongoengine==0.23.1
numpy==1.22.0
packaging==23.0
passlib==1.7.4
Pillow==9.4.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.6.0
pymongo==3.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
Werkzeug==2.0.2
WTForms==2.3.3
